fix: 修复duration字段类型不匹配
- 添加 duration::FLOAT 显式转换
- 解决 integer 与 double precision 类型不匹配

+++ supabase/migrations/20260111020000_optimize_video_tab_with_score.sql
@@ -38,7 +38,7 @@ BEGIN
     v.description::TEXT,
     v.cover_url::TEXT,
     v.play_url::TEXT,
-    v.duration,
+    v.duration::FLOAT,
     v.content_type::TEXT,
     v.tags,
     v.status::TEXT,
